Passing build; Code refactor
