+++ src/build.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 
 rm ../site/*
+rm main
 
-g++ -std=c++11 -DDEBUG -Wall -Wpedantic -Wshadow -Wextra -g -Og main.cpp -o main
+make
 
 ./main -a
