+++ src/classes/header-footer.cpp
@@ -1,11 +1,12 @@
 #include<string>
-#include "../Config/config.hpp"
+#include "../include/helpers.hpp"
+#include "../include/config.hpp"
 
 using namespace std;
-string static getHeader(Config *conf){
+string getHeader(Config *conf){
     return "<header>" + conf->header + "</header>"; 
 }
 
-string static getFooter(Config *conf){
+string getFooter(Config *conf){
    return "<footer><p>" + conf->footer + "</p></footer>"; 
 }
