+++ src/include/head.hpp
@@ -1,5 +1,5 @@
 #pragma once
 #include<string>
-#include"../Config/config.hpp"
+#include"config.hpp"
 
 std::string html_head(Config* conf, std::string text);
