+++ src/include/header-footer.hpp
@@ -1,0 +1,4 @@
+#include<string>
+#include"config.hpp"
+std::string getHeader(Config *conf);
+std::string getFooter(Config *conf);
