+++ src/include/helpers.hpp
@@ -4,3 +4,4 @@
 char getLower(char c);
 std::string toLowerCase(std::string text);
 void printError(int linenumber,std::string text);
+std::string parseLinks(std::string text, std::string path);
