+++ src/include/home.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include<string>
 
-#include "../Config/config.hpp"
-#include "../Category/category.hpp"
-#include"../../helpers/helpers.hpp"
+#include "config.hpp"
+#include "category.hpp"
+#include "helpers.hpp"
 void buildHome(Config *conf, Category *categories[], int categories_length, std::string path);
