+++ src/include/mizi.hpp
@@ -1,0 +1,2 @@
+#include<string>
+void createSite(std::string filename, std::string path);
